Fix: store transaction amount as NUMERIC to prevent precision drift

+++ app/schemas/transaction.py
@@ -1,5 +1,6 @@
 from pydantic import BaseModel, Field
 from typing import Optional
+from decimal import Decimal
 from datetime import datetime
 from app.models.transaction import TransactionStatus
 
@@ -7,14 +8,14 @@ class TransactionCreate(BaseModel):
     transaction_id: str = Field(..., min_length=1)
     source_account: str = Field(..., min_length=1)
     destination_account: str = Field(..., min_length=1)
-    amount: float = Field(..., gt=0)
+    amount: Decimal = Field(..., gt=0)
     currency: str = Field(..., min_length=1)
 
 class TransactionResponse(BaseModel):
     transaction_id: str
     source_account: str
     destination_account: str
-    amount: float
+    amount: Decimal
     currency: str
     status: TransactionStatus
     created_at: datetime
